Remove terminal.integrated.persistentSessionExperimentalSerializer
This now always uses the xterm.js-based serializer

Fixes #133021
