+++ src/vs/platform/terminal/common/terminalPlatformConfiguration.ts
@@ -366,12 +366,6 @@ const terminalPlatformConfiguration: IConfigurationNode = {
 			type: 'number',
 			default: 100
 		},
-		[TerminalSettingId.PersistentSessionExperimentalSerializer]: {
-			scope: ConfigurationScope.APPLICATION,
-			description: localize('terminal.integrated.persistentSessionExperimentalSerializer', "Whether to use a more efficient experimental approach for restoring the terminal's buffer. This setting requires a restart to take effect."),
-			type: 'boolean',
-			default: true
-		},
 		[TerminalSettingId.ShowLinkHover]: {
 			scope: ConfigurationScope.APPLICATION,
 			description: localize('terminal.integrated.showLinkHover', "Whether to show hovers for links in the terminal output."),
