+++ src/vs/platform/terminal/node/ptyHostService.ts
@@ -121,8 +121,7 @@ export class PtyHostService extends Disposable implements IPtyService {
 					VSCODE_VERBOSE_LOGGING: 'true', // transmit console logs from server to client,
 					VSCODE_RECONNECT_GRACE_TIME: this._reconnectConstants.graceTime,
 					VSCODE_RECONNECT_SHORT_GRACE_TIME: this._reconnectConstants.shortGraceTime,
-					VSCODE_RECONNECT_SCROLLBACK: this._reconnectConstants.scrollback,
-					VSCODE_RECONNECT_EXPERIMENTAL_SERIALIZATION: this._reconnectConstants.useExperimentalSerialization ? 1 : 0
+					VSCODE_RECONNECT_SCROLLBACK: this._reconnectConstants.scrollback
 				}
 			}
 		);
